Moved aways from the TcpClientPool abstraction and used a Registry instead.
After a brief exchange with some Elixir people on Slack,
it became clear to me that what I thought was a pool of tcp connections
was indeed, not a pool, in the sense of database connections.

What I did was more of a registry of processes. The problem with the former implementation,
was that when a process died, it was still present in the pool of client.
This problem is supposed to be solved with Elixir's registry. However, right now,
when a TCP connection is closed, the process is still well alive in the registry,
so it is something that I'll need to fix.

On top of that, I discovered that I could use Elixir's Registry as a kind of
pub-sub mechanism, where I could register several pairs of
TCP connections as well as their pids, under the same key,
and dispatch a message to all registered processes under the key.

It seems that this feature could be useful to built a system of private messages,
and rooms.